% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{a_marginals}
\alias{a_marginals}
\title{Marginal tables}
\usage{
a_marginals(A, am)
}
\arguments{
\item{A}{character matrix (the data)}

\item{am}{A list of variables (typically cliques or separators with RIP ordering)}
}
\description{
Returns a list with sparse marginal tables corresponding to variables in list \code{am}
}
