% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{adj_list}
\alias{adj_list}
\alias{adj_list.efs}
\title{Adjacency List}
\usage{
adj_list(x)

\method{adj_list}{efs}(x)
}
\arguments{
\item{x}{efs object}
}
\value{
An adjacency list
}
\description{
Extracts the adjacency list of a \code{efs} object
}
\seealso{
\code{\link{adj_matrix.efs}}
}
