% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{adj_matrix}
\alias{adj_matrix}
\alias{adj_matrix.efs}
\title{Adjacency Matrix}
\usage{
adj_matrix(x)

\method{adj_matrix}{efs}(x)
}
\arguments{
\item{x}{efs object}
}
\value{
An adjacency matrix
}
\description{
Extracts the adjacency matrix of a \code{efs} object
}
\seealso{
\code{\link{adj_list.efs}}
}
