% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{as_adj_mat}
\alias{as_adj_mat}
\title{Converts and adjacency list to an adjacency matrix}
\usage{
as_adj_mat(adj)
}
\arguments{
\item{adj}{Adjacency list}
}
\description{
Converts and adjacency list to an adjacency matrix
}
