% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{bws}
\alias{bws}
\title{Backward selection in decomposable graphical models}
\usage{
bws(df, adj, p = 0.5, ht = new.env(hash = TRUE), trace = TRUE,
  thres = 5)
}
\arguments{
\item{df}{data.frame}

\item{adj}{An adjacency list}

\item{p}{Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)}

\item{ht}{An environment (hashtable) containing precomputed entropies}

\item{trace}{Logical indicating whether or not to trace the procedure}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\value{
A bws object (a list) with values:
\describe{
\item{\code{G_adj}}{The fitted graph as an adjacency list.}
\item{\code{G_A}}{The fitted graph as an adjacency matrix.}
\item{\code{e}}{The edge that was just deleted and the stopping criteria attached to it.}
\item{\code{S}}{The minimal separator that now separated the nodes in \code{e} before deletion}
\item{\code{C}}{A list with the (maximal) cliques of the graph.}
\item{\code{ht}}{An updated version of the input argument \code{ht}.}
}
}
\description{
Backward selection in decomposable graphical models
}
\details{
See \code{\link{efs}} for details about \code{thres}.
}
\examples{
d <- tgp_dat[, 5:8]
G <- make_complete_graph(colnames(d))
bws(d, G)
}
\seealso{
\code{\link{bws_step}}, \code{\link{make_complete_graph}}, \code{\link{efs}}, \code{\link{efs_step}}, \code{\link{cl_tree}}
}
