% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_utils.R
\name{bws_class}
\alias{bws_class}
\title{Construct a bws object}
\usage{
bws_class(df, adj, ht = new.env(hash = TRUE))
}
\arguments{
\item{adj}{An adjacency list}

\item{ht}{An environment (hashtable) containing precomputed entropies}
}
\value{
A bws object. See \code{bws} for more details about the returning object.
}
\description{
A constructor for bws class
}
\examples{
d <- tgp_dat[, 5:8]
bws_class(d, make_complete_graph(colnames(d)))
}
\seealso{
\code{\link{bws_step}}, \code{\link{bws}}, \code{\link{make_complete_graph}}
}
