% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_utils.R
\name{bws_init}
\alias{bws_init}
\title{Construct a bws object}
\usage{
bws_init(df, adj, ht = new.env(hash = TRUE))
}
\arguments{
\item{df}{data.frame}

\item{adj}{An adjacency list}

\item{ht}{An environment (hashtable) containing precomputed entropies}
}
\value{
A bws object. See \code{bws} for more details about the returning object.
}
\description{
A constructor for bws class
}
\examples{
d   <- tgp_dat[, 5:8]
cg <- make_complete_graph(colnames(d))
bi <- bws_init(d, cg)
bws_step(d, bi)
}
\seealso{
\code{\link{bws_step}}, \code{\link{bws}}, \code{\link{make_complete_graph}}
}
