% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_utils.R
\name{bws_step}
\alias{bws_step}
\title{Stepwise backward selection in decomposable graphical models}
\usage{
bws_step(df, x, p = 0.5, thres = 5)
}
\arguments{
\item{df}{data.frame}

\item{x}{A bws object. Can be initialized with \code{bws_init}}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\value{
A bws object. See \code{bws} for details about the returning object.
}
\description{
Stepwise backward selection in decomposable graphical models
}
\details{
See \code{\link{efs}} for details about \code{thres}.
}
\examples{
d <- tgp_dat[, 5:8]
bws_step(d, bws_init(make_complete_graph(colnames(d))))
}
\references{
\url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100}
}
\seealso{
\code{\link{bws}}, \code{\link{make_complete_graph}}, \code{\link{efs}}, \code{\link{efs_step}}, \code{\link{cl_tree}}
}
