% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{cdf.outlier_model}
\alias{cdf.outlier_model}
\title{cdf}
\usage{
\method{cdf}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
The empirical cdf function of \code{T(Y)}
}
