% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{cdf}
\alias{cdf}
\alias{cdf.outlier_model}
\title{Empirical distribution function}
\usage{
cdf(x, ...)

\method{cdf}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatibility)}
}
\value{
The cumulative distribution of deviance test statistic of \code{x}
}
\description{
The empirical cdf of \code{T(Y)}
}
