% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{cl_tree}
\alias{cl_tree}
\title{Fitting a Chow-Liu tree interaction graph}
\usage{
cl_tree(df, wrap = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{wrap}{Logical indicating if the object should be converted to a \code{efs} object or not (can increase runtime significantly if FALSE)}
}
\description{
This functions uses Kruskals algorithm to fit a tree structure to variables in \code{df}
}
\examples{
cl_tree(tgp_dat[, 5:8])
}
\references{
\url{https://ieeexplore.ieee.org/abstract/document/1054142}
}
\seealso{
\code{\link{efs}}, \code{\link{efs_step}}, \code{\link{adj_list.efs}}, \code{\link{adj_matrix.efs}}
}
