% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{csptable}
\alias{csptable}
\title{Conditional sparse table}
\usage{
csptable(x, b)
}
\arguments{
\item{x}{A sparse table obtained from \code{sptable}}

\item{b}{A named vector of indicies for which the variables are fixed}
}
\description{
The names of \code{b} are the fixed values of the variables corresponding to the indicies
}
\details{
Returns a conditional sparse contingency table (also called a slice) for the variables in \code{x} as a vector.
}
\seealso{
\code{\link{sptable}}
}
