% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{cv}
\alias{cv}
\alias{cv.outlier_model}
\title{Critical value}
\usage{
cv(x, dz, alpha = 0.05, ...)

\method{cv}{outlier_model}(x, dz, alpha = 0.05, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{dz}{The deviance of the observation \code{z}.}

\item{alpha}{Significance level (between \code{0} and \code{1})}

\item{...}{Not used (for S3 compatability)}
}
\description{
Calculate the critical value for test statistic under \code{H_0}
}
\details{
The value \code{dz} can be obtained used the \code{deviance} function.
}
\seealso{
\code{\link{deviance}}
}
