% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{derma}
\alias{derma}
\title{Dermatology Database}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 358 rows and 35 columns.
}
\usage{
derma
}
\description{
We have removed 8 observations with missing values.
Data contains 12 clinical attributes and 21 histopathological attributes. The age
attribute has been discretized. The class variable has six levels; each
describing a skin disease.
}
\references{
\url{https://archive.ics.uci.edu/ml/datasets/dermatology}
}
\keyword{datasets}
