% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{deviance}
\alias{deviance}
\alias{deviance.outlier_model}
\title{Calculate deviance}
\usage{
deviance(x, y, ...)

\method{deviance}{outlier_model}(x, y, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{y}{An observation (name character vector)}

\item{...}{Not used (for S3 compatibility)}
}
\description{
This function calculates the affine value \code{T(y)} of \code{-2 log} likelihood-ratio statistic which is also called the deviance
}
