% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dgmsim.R
\name{dgm_sim_from_graph}
\alias{dgm_sim_from_graph}
\title{Simulate observations from a decomposable graphical model}
\usage{
dgm_sim_from_graph(g, lvls, nsim = 1000, cell_rate = 0.5)
}
\arguments{
\item{g}{An adjacency list}

\item{lvls}{Named list with levels of the discrete variables}

\item{nsim}{Number of simulations}

\item{cell_rate}{Control discrete cell probabilities}
}
\description{
Simulate observations from a decomposable graphical model
}
\examples{

\dontrun{

TBA

}
}
