% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digits}
\alias{digits}
\title{A data frame with observations corresponding to handwritten digits}
\format{An object of class \code{data.frame} with 5620 rows and 65 columns.}
\usage{
digits
}
\description{
The original data contained 32x32 bitmaps for each digit. Cell values are 1 if the grayscale color is more
towards black and zero otherwise. These bitmaps were then converted into blocks of 4x4 non-overlapping blocks
in which all the cell values were added (values ranging from 0 to 16).
The resulting matrix for an observations was then of dimension 8x8 (with each cell having values between 0 and 16).
}
\details{
In the final version of the data in molic, we have a data frame (of characters) with 5620 rows and 65 columns,
where the last column is the true class of the handwritten digit. All the remaining cells in the 64 columns was
converted to values in {a,b,c,d,e,f,g,h,i,j,k,l,m,n,o,p} corresponding the 16 possibilities. This is needed
in the outlier model. The to_single_chars() function can be exploited to achieve this for other data sets.
}
\references{
\href{https://archive.ics.uci.edu/ml/datasets/optical+recognition+of+handwritten+digits}{Optical Recognition of Handwritten Digits}
}
\keyword{datasets}
