% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{efs.print}
\alias{efs.print}
\title{Print}
\usage{
efs.print(x, ...)
}
\arguments{
\item{x}{A \code{efs} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{efs} objects
}
