% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{efs}
\alias{efs}
\title{Efficient Forward Selection in Decomposable Graphical Models}
\usage{
efs(df, x = efs_init(df), d = 3, trace = TRUE, mdl_type = "mdl1")
}
\arguments{
\item{df}{Dataframe}

\item{x}{An efs object}

\item{d}{Number of bits to encode a single parameter}

\item{trace}{Logical indidcating whether or not to trace the procedure}

\item{mdl_type}{Temporary parameter}
}
\description{
Efficient Forward Selection in Decomposable Graphical Models
}
