% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{efs_adj_list.efs}
\alias{efs_adj_list.efs}
\title{Adjacency List}
\usage{
\method{efs_adj_list}{efs}(x)
}
\arguments{
\item{x}{efs object}
}
\value{
An adjacency list
}
\description{
Extracts the adjacency list of a \code{efs} object
}
\examples{
## TBA

}
\seealso{
\code{\link{efs_adj_matrix.efs}}
}
