% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{efs_adj_matrix.efs}
\alias{efs_adj_matrix.efs}
\title{Adjacency Matrix}
\usage{
\method{efs_adj_matrix}{efs}(x)
}
\arguments{
\item{x}{efs object}
}
\value{
An adjacency matrix
}
\description{
Extracts the adjacency matrix of a \code{efs} object
}
\examples{
## TBA

}
\seealso{
\code{\link{efs_adj_list.efs}}
}
