% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_utils.R
\name{efs_init}
\alias{efs_init}
\title{Initialize}
\usage{
efs_init(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
A efs object
}
\description{
Initialize The EFS algorithm
}
\examples{
efs_init(tgp_dat[, 5:8])
}
\references{
\url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100}
}
\seealso{
\code{\link{cl_tree}}, \code{\link{efs}}, \code{\link{efs_step}}, \code{\link{adj_list.efs}}, \code{\link{adj_matrix.efs}}
}
