% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_utils.R
\name{efs_step}
\alias{efs_step}
\title{EFS step}
\usage{
efs_step(df, x, thres = 5)
}
\arguments{
\item{df}{Dataframe}

\item{x}{An efs object}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy}
}
\description{
EFS step
}
