% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_graph.R
\name{fit_graph}
\alias{fit_graph}
\title{Fit a decomposable graphical model}
\usage{
fit_graph(df, type = "fwd", adj = NULL, q = 0.5, trace = TRUE,
  thres = 5, wrap = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{type}{Character ("fwd", "bwd", "tree")}

\item{adj}{Adjacency list of a decomposable graph}

\item{q}{Penalty term in the stopping criterion (\code{0} = AIC and \code{1} = BIC)}

\item{trace}{Logical indicating whether or not to trace the procedure}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy.}

\item{wrap}{logical specifying if the result of a run with type = "tree" should be converted to a "fwd" object}
}
\value{
A \code{gengraph} object with values:
}
\description{
A generic method for structure learning in decomposable graphical models
}
\examples{
\dontrun{
library(dplyr)

## All handwritten digits that have true class equal to a "1".
d <- digits \%>\%
  filter(class == "1") \%>\%
  select(-class) 

g <- fit_graph(d)
print(g)
plot(g, vertex.size = 1)
adj_mat(g)
adj_lst(g)
}
}
\references{
\url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100}
}
\seealso{
\code{\link{adj_lst.gengraph}}, \code{\link{adj_mat.gengraph}}, \code{\link{walk.fwd}}, \code{\link{walk.bwd}}, \code{\link{gengraph}}
}
