% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_api_fit_tree.R
\name{fit_tree}
\alias{fit_tree}
\title{Fitting a Chow-Liu tree interaction graph}
\usage{
fit_tree(df, wrap = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{wrap}{Logical indicating if the object should be converted to a \code{fwd} object or not (can increase runtime significantly if FALSE)}
}
\description{
This functions uses Kruskals algorithm to fit a tree structure to variables in \code{df}
}
\examples{
G <- cl_tree(tgp_dat[, 5:8])
print(G)
plot(G)
}
\references{
\url{https://ieeexplore.ieee.org/abstract/document/1054142}
}
\seealso{
\code{\link{efs}}, \code{\link{efs_step}}, \code{\link{bws}}, \code{\link{bws_step}},\code{\link{adj_list.efs}}, \code{\link{adj_matrix.efs}}
}
