% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gengraph.R
\name{gengraph}
\alias{gengraph}
\title{Fit a decomposable graphical model}
\usage{
gengraph(df, type = "fwd", adj = NULL, q = 0.5, ...)
}
\arguments{
\item{df}{data.frame}

\item{type}{character ("fwd", "bwd", "tree")}

\item{adj}{A userspecified adjacency list}

\item{q}{Penalty term in the stopping criterion (\code{0} = AIC and \code{1} = BIC)}

\item{...}{Not used (for extendibility)}
}
\value{
A \code{gengraph} object
}
\description{
A generic method for structure learning in decomposable graphical models
}
\examples{
d <- digits[1:100, 1:5]
gengraph(d)
}
\seealso{
\code{\link{adj_lst.gengraph}}, \code{\link{adj_mat.gengraph}}, \code{\link{fit_graph}}, \code{\link{walk.fwd}}, \code{\link{walk.bwd}}
}
