% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_api_jt.R
\name{get_cliques}
\alias{get_cliques}
\alias{get_cliques.jt}
\title{Return the cliques of a junction tree}
\usage{
get_cliques(x)

\method{get_cliques}{jt}(x)
}
\arguments{
\item{x}{A junction tree object, \code{jt}.}
}
\description{
Return the cliques of a junction tree
}
\seealso{
\code{\link{jt}}
}
