% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{is_decomposable}
\alias{is_decomposable}
\title{A test for decomposability in undirected graphs}
\usage{
is_decomposable(adj)
}
\arguments{
\item{adj}{Adjacency list of an undirected graph}
}
\description{
This function returns \code{TRUE} if the graph is decomposable and \code{FALSE} otherwise
}
