% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_jt.R
\name{jt}
\alias{jt}
\title{Junction Tree}
\usage{
jt(g, data, evidence = NULL, flow = "sum", propagate = TRUE, validate = TRUE)
}
\arguments{
\item{g}{Either a decomposable undirected graph as a named list or a \code{gengraph} object or an \code{igraph} DAG}

\item{data}{A \code{data.frame} of characters. Note: all values must only be a single character. See details.}

\item{evidence}{A named vector}

\item{flow}{Character. Either "sum" or "max"}

\item{propagate}{Logical}

\item{validate}{Logical. See details.}
}
\value{
A \code{jt} object
}
\description{
Construction of a junction tree and compilation of the network
}
\details{
It is assumed that all values in \code{data}, for all variables,
are represented as a single character. If \code{validate} is \code{TRUE} this is checked.
If values are not single characters, one may exploit the \code{to_single_chars} function.
}
\examples{

# Setting up the network
# ----------------------
library(igraph)
set.seed(7)
g <- make_tree(7)
g <- g + edge(4, 7)
g <- g + edge(5, 6)
g <- add_vertices(g, 3)
g <- g + edge(9, 10)
g <- set.vertex.attribute(g, "name", value = letters[1:10])
plot(g)

# Some arbitrary data
d <- tgp_dat[1:500, 9:18]
colnames(d) <- letters[1:vcount(g)]
# -----------------------

# Example 1: sum-flow without evidence
jt1 <- jt(g, d)
plot(jt1)
print(jt1)
query_belief(jt1, c("a", "e", "g"))
query_belief(jt1, c("a", "d", "g"), type = "joint")

# Example 2: sum-flow with evidence
e2  <- c(a = "T", b = "C")
jt2 <- jt(g, d, e2)
query_belief(jt2, c("a", "e", "g"))
query_belief(jt2, c("a", "d", "g"), type = "joint")
# Notice that, the configuration "TTT" has changed
# dramatically as a consequence of the evidence

# Example 3: max-flow without evidence
jt3 <- jt(g, d, flow = "max")
mpe(jt3)

# Example 4: max-flow with evidence
e4  <- c(a = "T", b = "C", c = "A", j = "A", g = "T")
jt4 <- jt(g, d, e4, flow = "max")
mpe(jt4)

# Notice, that the value of "d" has changed to "C" compared
# to max(jt3) where the value was "T" as a consequence
# of the new evidence
}
\seealso{
\code{\link{query_belief}}, \code{\link{mpe}}, \code{\link{get_cliques}}
}
