% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_misc.R
\name{make_complete_graph}
\alias{make_complete_graph}
\title{Make a complete graph}
\usage{
make_complete_graph(nodes)
}
\arguments{
\item{nodes}{A character vector containing the nodes to be used in the graph}
}
\description{
A helper function to make an adjacency list corresponding to a complete graph
}
\examples{
d  <- digits[, 5:8]
cg <- make_complete_graph(colnames(d))
}
