% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_api_jt.R
\name{max_config}
\alias{max_config}
\alias{max_config.jt}
\title{Max Configuration}
\usage{
max_config(x)

\method{max_config}{jt}(x)
}
\arguments{
\item{x}{A junction tree object, \code{jt}, with max-flow.}
}
\description{
Returns the most probable configuration given the evidence entered in the junction tree
}
\examples{
# See the 'jt' function
}
\seealso{
\code{\link{jt}}
}
