% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{mean.outlier_model}
\alias{mean.outlier_model}
\title{Mean}
\usage{
\method{mean}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatibility)}
}
\value{
The mean of the deviance test statistic of \code{x}
}
\description{
Estimated mean of deviance statistic \code{T(Y)}
}
