% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molic.R
\docType{package}
\name{molic-package}
\alias{molic}
\alias{molic-package}
\title{molic: Multivariate Ouliterdetection In Contingency Tables}
\description{
Outlier detection in, possibly highdimensional, categorical data following
Mads Lindskou et al. (2019) <doi:10.1111/sjos.12407>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlindsk/molic}
  \item Report bugs at \url{https://github.com/mlindsk/molic/issues}
}

}
\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
