% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{outlier_model}
\alias{outlier_model}
\title{Outlier tests in contingency tables using decomposable graphical models}
\usage{
outlier_model(df, adj, nsim = 1000, ncores = 1, meta_name = "")
}
\arguments{
\item{df}{Data frame}

\item{adj}{Adjacency list of a decomposable graph}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{meta_name}{A meta name to keep track of different outlier models}
}
\description{
Outlier tests in contingency tables using decomposable graphical models
}
\details{
It is assumed that all cell values in \code{df}, for all variables,
are represented as a single character. No check for this is done in the
current version. If cell values are not single characters, one may
exploit \code{letters} and \code{LETTERS} e.g.
}
