% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{p_val.outlier_model}
\alias{p_val.outlier_model}
\title{p_val}
\usage{
\method{p_val}{outlier_model}(x, ty_new, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{ty_new}{The transformed value T(y_new) obtained from function \code{TY}}

\item{...}{Not used (for S3 compatability)}
}
\description{
Calculate the p value for obtaining ty_new under \code{H_0}
}
