% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{p_val}
\alias{p_val}
\alias{p_val.outlier_model}
\title{P-value}
\usage{
p_val(x, dy, ...)

\method{p_val}{outlier_model}(x, dy, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{dy}{The deviance of the observation \code{y}.}

\item{...}{Not used (for S3 compatability)}
}
\description{
Calculate the p-value for obtaining \code{ty_new} under \code{H_0}
}
\details{
The value \code{ty_new} can be obtained used the \code{deviance} function.
}
\seealso{
\code{\link{deviance}}
}
