% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{parray}
\alias{parray}
\title{Conditional probability table}
\usage{
parray(x, y = NULL)
}
\arguments{
\item{x}{A sparse table obtained from \code{sptable}}

\item{y}{Character vector with variables to condition on}
}
\description{
Returns a sparse conditional probability table conditioned on variables in \code{b}.
}
\details{
If \code{y} is \code{NULL}, \code{x} is just converted to a \code{sptable} with no conditioning variables, i.e. the marginal table.
}
\examples{
sp  <- sptable(as.matrix(digits[, 1:3]))
psp <- parray(sp, c("V1", "V2"))
sum(psp[1:length(letters[1:17])]) # V3 has 17 levels.
}
\seealso{
\code{\link{sptable}}, \code{\link{slice_sptable}}
}
