% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{plot.bws}
\alias{plot.bws}
\title{Plot bws}
\usage{
\method{plot}{bws}(x, ...)
}
\arguments{
\item{x}{A \code{bws} object}

\item{...}{Extra arguments. See the igraph package}
}
\description{
A wrapper around igraphs plot method for \code{bws} objects
}
