% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{plot.efs}
\alias{plot.efs}
\title{Plot efs}
\usage{
\method{plot}{efs}(x, ...)
}
\arguments{
\item{x}{A \code{efs} object}

\item{...}{Extra arguments. See the igraph package}
}
\description{
A wrapper around igraphs plot method for \code{efs} objects
}
