% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{plot.gengraph}
\alias{plot.gengraph}
\title{Plot}
\usage{
\method{plot}{gengraph}(x, ...)
}
\arguments{
\item{x}{A \code{gengraph} object}

\item{...}{Extra arguments. See the igraph package}
}
\description{
A wrapper around igraphs plot method for \code{gengraph} objects
}
\examples{
d <- subset(digits, class == "1")[, 20:30]
g <- gengraph(d, type = "fwd")
plot(g)
G <- fit_graph(d)
plot(G)
}
