% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{plot.multiple_models}
\alias{plot.multiple_models}
\title{Plot Deviance of Multiple Models}
\usage{
\method{plot}{multiple_models}(x, sig_col = "#FF0000A0", ...)
}
\arguments{
\item{x}{A \code{multiple_models} object returned from a called to \code{fit_multiple_models}}

\item{sig_col}{Color of the significance level area (default is red)}

\item{...}{Extra arguments. See details.}
}
\description{
A plot method to show the the approximated deviance distribution of multiple models
}
\details{
No extra functionalities are implemented for \code{...} yet.
}
\examples{
# TBA
}
