% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{plot.outlier}
\alias{plot.outlier}
\title{Plot Deviance}
\usage{
\method{plot}{outlier}(x, sig_col = "#FF0000A0", ...)
}
\arguments{
\item{x}{An object returned from \code{fit_outlier}}

\item{sig_col}{Color of the significance level area (default is red)}

\item{...}{Extra arguments; see details.}
}
\description{
A plot method to show the the approximated deviance distribution
}
\details{
The dotted line represents the observed deviance of the observation under the hypothesis
and the colored (red is default) area under the graph represents the significance level.
Thus, if the dotted line is to the left of the colored area, the hypothesis that the observation
is an outlier cannot be rejected. Notice however, if there is no dotted line, this simply means,
that the observed deviance is larger than all values and it would disturb the plot if included.

No extra arguments \code{...} are implement at the moment.
}
\examples{
# TBA
1L
}
