% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{plot.outlier_model}
\alias{plot.outlier_model}
\title{Plot Deviance}
\usage{
\method{plot}{outlier_model}(x, type = "base", ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{type}{either "base" og "ggplot"}

\item{...}{Extra arguments; see details.}
}
\description{
A plot method to show the the approximated deviance distribution
}
\details{
\itemize{
  \item TBA 
}
}
\examples{
# TBA
}
