% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{pmf}
\alias{pmf}
\alias{pmf.outlier_model}
\title{Plot of pmf}
\usage{
pmf(x, ...)

\method{pmf}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A plot method to show the pmf of the approximated pmf of \code{T(Y)}
}
