% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R
\name{print.bws}
\alias{print.bws}
\title{Print bws}
\usage{
\method{print}{bws}(x, ...)
}
\arguments{
\item{x}{A \code{efs} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{bws} objects
}
