% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{print.csptable}
\alias{print.csptable}
\title{Print csptable}
\usage{
\method{print}{csptable}(x, ...)
}
\arguments{
\item{x}{A \code{csptable} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{csptable} objects
}
