% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{print.efs}
\alias{print.efs}
\title{Print}
\usage{
\method{print}{efs}(x, ...)
}
\arguments{
\item{x}{A \code{efs} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{efs} objects
}
