% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{print.outlier}
\alias{print.outlier}
\title{Print outlier}
\usage{
\method{print}{outlier}(x, ...)
}
\arguments{
\item{x}{A \code{outlier} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{outlier} objects
}
