% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{print.outlier_model}
\alias{print.outlier_model}
\title{Print outlier model}
\usage{
\method{print}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatability)}
}
\value{
No return value, called for side effects
}
\description{
A print method for \code{outlier_model} objects
}
