% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print.tree}
\alias{print.tree}
\title{Print tree}
\usage{
\method{print}{tree}(x, ...)
}
\arguments{
\item{x}{A \code{tree} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{tree} objects
}
