% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{pval}
\alias{pval}
\alias{pval.outlier_model}
\title{P-value}
\usage{
pval(x, dz, ...)

\method{pval}{outlier_model}(x, dz, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{dz}{The deviance of the observation \code{z}.}

\item{...}{Not used (for S3 compatibility)}
}
\value{
The p-value of deviance test statistic of \code{x}
}
\description{
Calculate the p-value for obtaining \code{ty_new} under \code{H_0}
}
\details{
The value \code{dz} can be obtained used the \code{deviance} function.
}
\seealso{
\code{\link{deviance}}
}
