% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rip}
\alias{rip}
\title{Runnining Intersection Property}
\usage{
rip(adj)
}
\arguments{
\item{adj}{A named adjacency list}
}
\description{
Given a decomposable graph, this functions finds a perfect numbering on the vertices using maximum cardinality search, and hereafter returns a list with two elements: "C" - A RIP-ordering of the cliques and "S" - A RIP ordering of the separators.
}
