% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{sim_TY}
\alias{sim_TY}
\title{Simulation of TY}
\usage{
sim_TY(A, C_marginals, S_marginals, nsim = 1000, ncores = 1)
}
\arguments{
\item{A}{Character Matrix (data)}

\item{C_marginals}{Clique marginal tables}

\item{S_marginals}{Separator marginal tables}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}
}
\description{
This function simulates observations Ty in order to obtain the approximated density of TY
}
