% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{sim_Y}
\alias{sim_Y}
\title{Simulation of cells in contingency tables}
\usage{
sim_Y(A, C_marginals, S_marginals, nsim = 1000, ncores = 1)
}
\arguments{
\item{A}{Character Matrix (data)}

\item{C_marginals}{Clique marginal tables}

\item{S_marginals}{Separator marginal tables}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}
}
\description{
This function simulates cells in a contingency table based on a decomposable graphical model
}
