% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{slice_sptable}
\alias{slice_sptable}
\title{Slice of a sparse table}
\usage{
slice_sptable(x, b)
}
\arguments{
\item{x}{A \code{sptable} object}

\item{b}{A named vector of indicies for which the variables are fixed}
}
\description{
The names of \code{b} are the fixed values of the variables corresponding to the indicies
}
\details{
Returns the b-slice of a conditional sparse contingency table for the variables in \code{x} as a vector.
}
\examples{
sp <- sptable(as.matrix(digits[, 1:3]))
# print(sp)
slice_sptable(sp, c(a = 2))
}
\seealso{
\code{\link{sptable}}
}
