% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{sptable}
\alias{sptable}
\title{Sparse table}
\usage{
sptable(x)
}
\arguments{
\item{x}{matrix}
}
\description{
Returns a sparse contingency table for the variables in \code{x} as a vector .
}
\seealso{
\code{\link{csptable}}
}
