% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_step_bwd.R
\name{step.bwd}
\alias{step.bwd}
\title{Stepwise backward selection}
\usage{
\method{step}{bwd}(x, df, q = 0.5, thres = 5)
}
\arguments{
\item{x}{\code{gengraph}}

\item{df}{data.frame}

\item{q}{Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\value{
A \code{bwd} object (a subclass of \code{gengraph})
}
\description{
Stepwise backward selection in decomposable graphical models
}
\details{
A \code{bwd} object can be created using the \code{gengraph} constructor with \code{type = "bwd"}
}
\seealso{
\code{\link{fit_graph}}, \code{\link{step.fwd}}, \code{\link{gengraph}}
}
