% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tgp_dat}
\alias{tgp_dat}
\title{A data frame with genetic data from the 1000 genomes project}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5008 rows and 304 columns.
}
\usage{
tgp_dat
}
\description{
The data consists of \code{2504} DNA profiles, each genotyped on 304 SNPs (binary variables).
The data frame has \code{5008} rows, since each profile has two copies.
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/26432245}{1000 Genomes Project}
}
\keyword{datasets}
