% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tgp_haps}
\alias{tgp_haps}
\title{A named list of character vectors.}
\format{
An object of class \code{list} of length 109.
}
\usage{
tgp_haps
}
\description{
Every element in the list is a character vector that forms a haplotype
from the 1000 genomes project. If the list is unlisted, it should correspond
to colnames of tgp_dat. In other words, tgp_haps is a "haplotype-grouping"
of the variables in tgp_dat.
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/26432245}{1000 Genomes Project}
}
\keyword{datasets}
