% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tgp_meta}
\alias{tgp_meta}
\title{A data frame with meta information of the data in \code{tgp_haps}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 130 rows and 5 columns.}
\usage{
tgp_meta
}
\description{
The data frame holds information about haplotypes and on which chromosome
these are located.
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/26432245}{1000 Genomes Project}
}
\keyword{datasets}
