% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_misc.R
\name{to_single_chars}
\alias{to_single_chars}
\title{Convert discrete values into a single character representation}
\usage{
to_single_chars(x)
}
\arguments{
\item{x}{Data frame or matrix of characters}
}
\description{
Convert all values in a data frame or matrix of characters to a single
character representation
}
\examples{
d <- data.frame(x = c("11", "2"), y = c("2", "11"))
to_single_chars(d)
}
