% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TY}
\alias{TY}
\title{T(y)}
\usage{
TY(y, C_marginals, S_marginals)
}
\arguments{
\item{y}{A named (according to data) vector}

\item{C_marginals}{Clique marginal tables (use a_marginals function)}

\item{S_marginals}{Separator marginal tables (use a_marginals function)}
}
\description{
Calculate the affine value T(y) of -2log likelihood-ratio statistic
}
