% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{variance.outlier_model}
\alias{variance.outlier_model}
\title{Variance}
\usage{
\method{variance}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
Estimated variance of \code{T(Y)}
}
