% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{walk}
\alias{walk}
\title{Stepwise model selection}
\usage{
walk(x, df, q, thres)
}
\arguments{
\item{x}{\code{fwd} or \code{bwd} objects}

\item{df}{data.frame}

\item{q}{Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\description{
Stepwise model selection in decomposable graphical models
}
\details{
A \code{bwd} (or \code{fwd}) object can be created using the \code{gengraph} constructor with \code{type = "bwd"}
}
\examples{
d <- subset(digits, class == "1")[, 30:35]
gb <- gengraph(d, type = "bwd")
gf <- gengraph(d, type = "fwd")
wb <- walk(gb, d)
wf <- walk(gf, d)
# plot(wf)
# adj_mat(wf)
# adj_lst(wf)
}
\seealso{
\code{\link{fit_graph}}, \code{\link{walk.fwd}}, \code{\link{gengraph}}
}
