% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R, R/efs.R
\name{adj_list.bws}
\alias{adj_list.bws}
\alias{adj_list}
\alias{adj_list.efs}
\title{Adjacency List}
\usage{
\method{adj_list}{bws}(x)

adj_list(x)

\method{adj_list}{efs}(x)
}
\arguments{
\item{x}{\code{efs} or \code{bws} object}
}
\value{
An adjacency list
}
\description{
Extracts the adjacency list of a \code{efs} or \code{bws} object
}
