% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws.R, R/efs.R
\name{adj_matrix.bws}
\alias{adj_matrix.bws}
\alias{adj_matrix}
\alias{adj_matrix.efs}
\title{Adjacency Matrix}
\usage{
\method{adj_matrix}{bws}(x)

adj_matrix(x)

\method{adj_matrix}{efs}(x)
}
\arguments{
\item{x}{\code{efs} or \code{bws} object}
}
\value{
An adjacency matrix
}
\description{
Extracts the adjacency matrix of a \code{efs} or \code{bws} object
}
