% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_utils.R
\name{bws_init}
\alias{bws_init}
\title{Initialize a bws object}
\usage{
bws_init(adj, ht = new.env(hash = TRUE))
}
\arguments{
\item{adj}{An adjacency list}

\item{ht}{An environment (hashtable) containing precomputed entropies}
}
\value{
A bws object. See \code{bws} for more details about the returning object.
}
\description{
Initialize a bws object
}
\examples{
d <- tgp_dat[, 5:8]
bws_init(make_complete_graph(colnames(d)))
}
\seealso{
\code{\link{bws_step}}, \code{\link{bws}}, \code{\link{make_complete_graph}}
}
