% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bws_utils.R
\name{bws_step}
\alias{bws_step}
\title{Stepwise backward selection in decomposable graphical models}
\usage{
bws_step(df, x, p = 0.5, thres = 5)
}
\arguments{
\item{df}{data.frame}

\item{x}{A bws object. Can be created explicitly with \code{bws_init}}

\item{p}{Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\value{
A bws object. See \code{bws} for details about the returning object.
}
\description{
Stepwise backward selection in decomposable graphical models
}
\details{
See \code{\link{efs}} for details about \code{thres}.
}
\examples{
d  <- tgp_dat[, 5:8]
cg <- make_complete_graph(colnames(d))
bi <- bws_init(d, cg)
bws_step(d, bi)
}
\seealso{
\code{\link{bws}}, \code{\link{make_complete_graph}}, \code{\link{efs}}, \code{\link{efs_step}}, \code{\link{cl_tree}}
}
