% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{cl_tree}
\alias{cl_tree}
\title{Fitting a Chow-Liu tree interaction graph}
\usage{
cl_tree(df, wrap = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{wrap}{Logical indicating if the object should be converted to a \code{efs} object or not (can increase runtime significantly if FALSE)}
}
\description{
This functions uses Kruskals algorithm to fit a tree structure to variables in \code{df}
}
\references{
\url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100}
}
\seealso{
\code{\link{efs}}, \code{\link{efs_step}}, \code{\link{efs_adj_list.efs}}, \code{\link{efs_adj_matrix.efs}}
}
