% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{csptable}
\alias{csptable}
\title{Conditional sparse table}
\usage{
csptable(x, y, parray = TRUE)
}
\arguments{
\item{x}{A sparse table obtained from \code{sptable}}

\item{y}{Character vector with variables to condition on}

\item{parray}{Logical if the cell values should be converted to probabilities}
}
\description{
Returns a conditional sparse contingency table conditioned on variables in \code{b}.
}
\details{
If \code{y} is \code{NULL}, \code{x} is just converted to a \code{csptable} with no conditioning variables, i.e. the marginal table.
}
\examples{
sp <- sptable(as.matrix(digits[, 1:3]))
sp
length(sp)
csp <- csptable(sp, c("V3"))
csp
length(csp)
}
\seealso{
\code{\link{sptable}}, \code{\link{slice_sptable}}
}
