% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{dgm_sim}
\alias{dgm_sim}
\title{Simulate observations from a decomposable graphical model}
\usage{
dgm_sim(A, adj, nsim = 1000, ncores = 1)
}
\arguments{
\item{A}{Character Matrix (data)}

\item{adj}{Adjacency list of a decomposable graph}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}
}
\value{
This function returns a matrix of dimension \code{nsim x ncol(A)} where each row correspond to a simulated observation from a DGM represented by \code{adj}.
}
\description{
This function simulates observations from a DGM
}
\examples{

\dontrun{

library(dplyr)

d <- digits %>%
  filter(class == "1") %>%
  select(-class)
G <- adj_lst(fit_graph(d))
dgm_sim(as.matrix(d), G, nsim = 10)

}
}
