% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{efs}
\alias{efs}
\title{efs}
\usage{
efs(df, x = efs_init(df), trace = TRUE, stop_crit = "mdl1",
  thres = 5)
}
\arguments{
\item{df}{Dataframe}

\item{x}{An efs object}

\item{trace}{Logical indidcating whether or not to trace the procedure}

\item{stop_crit}{Stopping criterion (mdl1, mdl2, aic or bic)}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy}
}
\description{
Efficient forward felection in decomposable graphical models
}
