% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs.R
\name{efs}
\alias{efs}
\title{Efficient forward selection in decomposable graphical models}
\usage{
efs(df, x = efs_init(df), p = 0.5, trace = TRUE, thres = 5)
}
\arguments{
\item{df}{data.frame}

\item{x}{A efs object}

\item{p}{Penalty term in the stopping criterion  (\code{0} = AIC and \code{1} = BIC)}

\item{trace}{Logical indicating whether or not to trace the procedure}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\value{
A efs object
}
\description{
Efficient forward selection in decomposable graphical models
}
\examples{
efs(tgp_dat[, 5:8])
}
\references{
\url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100}
}
\seealso{
\code{\link{cl_tree}}, \code{\link{efs_step}}, \code{\link{adj_list.efs}}, \code{\link{adj_matrix.efs}}
}
