% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efs_utils.R
\name{efs_step}
\alias{efs_step}
\title{Stepwise efficient forward selection in decomposable graphical models}
\usage{
efs_step(df, x, thres = 5)
}
\arguments{
\item{df}{data.frame}

\item{x}{A efs object}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\value{
A efs object
}
\description{
Stepwise efficient forward selection in decomposable graphical models
}
\details{
See \code{\link{efs}} for details about \code{thres}
}
\examples{
x <- tgp_dat[, 5:8]
efs_step(x, efs_init(x))
}
\references{
\url{https://arxiv.org/abs/1301.2267}, \url{https://doi.org/10.1109/ictai.2004.100}
}
\seealso{
\code{\link{efs}}, \code{\link{bws}}, \code{\link{bws_step}}, \code{\link{make_complete_graph}}, \code{\link{cl_tree}}
}
