% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{fit_mixed_outlier}
\alias{fit_mixed_outlier}
\title{Mixed Outlier Test}
\usage{
fit_mixed_outlier(m1, m2, z1, z2, alpha = 0.05)
}
\arguments{
\item{m1}{A \code{outlier_model} object}

\item{m2}{A \code{outlier_model} object}

\item{z1}{Named vector (same names as \code{colnames(A)})}

\item{z2}{Named vector (same names as \code{colnames(A)})}

\item{alpha}{The significance level}
}
\description{
A convinient function for outlier detection with mixed information
}
\examples{
# See the package vignette "Outlier Detection in Genetic Data"
}
