% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{fit_mixed_outlier}
\alias{fit_mixed_outlier}
\title{Mixed Outlier Test}
\usage{
fit_mixed_outlier(m1, m2)
}
\arguments{
\item{m1}{An object returned from \code{fit_outlier}}

\item{m2}{An object returned from \code{fit_outlier}}
}
\description{
A convenient function for outlier detection with mixed information
}
\examples{
\dontrun{

# See the package vignette "Outlier Detection in Genetic Data"
for an introduction to the genetic data used here

library(dplyr)

# The components - here microhaplotypes
haps <- tgp_haps

# All the Europeans
eur  <- tgp_dat \%>\%
  as_tibble() \%>\%
  filter(pop_meta == "EUR")

# Extracting the two databases for each copy of the chromosomes
eur_a <- eur \%>\%
  filter(grepl("a$", sample_name)) \%>\%
  select(-c(1:2))

eur_b <- eur \%>\%
  filter(grepl("b$", sample_name)) \%>\%
  select(-c(1:2))

mat_eur_a <- eur_a \%>\% select(-c(1:2)) \%>\% as.matrix()
mat_eur_b <- eur_b \%>\% select(-c(1:2)) \%>\% as.matrix()

# Fitting the interaction graphs on the EUR data 
ga <- fit_components(eur_a, comp = haps, trace = FALSE, as_gen = TRUE)
gb <- fit_components(eur_b, comp = haps, trace = FALSE, as_gen = TRUE)

# as_gen = TRUE converts the graph to a gengraph,
# the main object for graphs in the molic package
# This enables plotting and print info of the graph:
# print(ga)
# plot(ga, vertex.size = 1)

# Testing if an American is an outlier in Europe

amr <- tgp_dat \%>\%
  as_tibble() \%>\%
  filter(pop_meta == "AMR")


z1  <- amr \%>\%
  filter(grepl("a$", sample_name)) \%>\% 
  select(unname(unlist(haps))) \%>\%
  slice(1) \%>\%
  unlist()

z2  <- amr \%>\%
  filter(grepl("b$", sample_name)) \%>\% 
  select(unname(unlist(haps))) \%>\%
  slice(1) \%>\%
  unlist()


m1 <- fit_outlier(mat_eur_a, z1, ga) # consider using more cores (ncores argument)
m2 <- fit_outlier(mat_eur_a, z2, gb) # consider using more cores (ncores argument)
m  <- fit_mixed_outlier(m1,m2)
plot(m)
}
}
