% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{fit_multiple_models}
\alias{fit_multiple_models}
\title{Fit Multiple Models}
\usage{
fit_multiple_models(
  A,
  z,
  response,
  alpha = 0.05,
  type = "fwd",
  q = 0.5,
  nsim = 10000,
  ncores = 1,
  trace = TRUE,
  validate = TRUE
)
}
\arguments{
\item{A}{A data frame (data) without the new observation \code{z} appended}

\item{z}{Named vector (same names as \code{colnames(A)} but without the class variable)}

\item{response}{A character with the name of the class variable of interest}

\item{alpha}{The significance level}

\item{type}{Character ("fwd", "bwd", "tree" or "tfwd") - the type of interaction graph to be used}

\item{q}{Penalty term in the stopping criterion when fitting the interaction graph (\code{0} = AIC and \code{1} = BIC)}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{trace}{Logical indicating whether or not to trace the procedure}

\item{validate}{Logical. If true, it checks if \code{A} has only single character values and converts it if not.}
}
\description{
A convinient wrapper function to conduct multiple tests for a single observation
}
\examples{
# TBA
1L
}
