% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{fit_outlier}
\alias{fit_outlier}
\title{Outlier Test}
\usage{
fit_outlier(A, z, adj, type = "fwd", alpha = 0.05, nsim = 5000,
  ncores = 1, validate = TRUE)
}
\arguments{
\item{A}{Character Matrix (data)}

\item{z}{Named vector (same names as \code{colnames(A)})}

\item{adj}{Adjacency list of a decomposable graph}

\item{type}{Character ("fwd", "bwd", "tree")}

\item{alpha}{The significance level}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{validate}{Logical. If true, it checks if \code{A} has only single character values and converts it if not.}
}
\description{
A convinient wrapper around the \code{outlier_model} for outlier detection
}
\examples{
# All handwritten digits that have true class equal to a "1".
d <- subset(digits, class == "1")
d[, "class"] <- NULL

# A handwritten digit with true class equal to "7"
zd <- subset(digits, class == "7")
z  <- unlist(zd[1, 1:(ncol(zd)-1)])

# Fit an interaction graph
G <- fit_graph(d)
plot(G, vertex.size = 1)

# Test if z is an outlier
M <- fit_outlier(as.matrix(d), z, adj_lst(G))
M
pmf(M)
}
