% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{fit_outlier}
\alias{fit_outlier}
\title{Fit Outlier}
\usage{
fit_outlier(
  A,
  z,
  alpha = 0.05,
  type = "fwd",
  q = 0.5,
  comp = NULL,
  nsim = 10000,
  ncores = 1,
  trace = FALSE,
  validate = TRUE
)
}
\arguments{
\item{A}{Character Matrix (data) - without the new observation \code{z} appended}

\item{z}{Named vector (same names as \code{colnames(A)})}

\item{alpha}{The significance level}

\item{type}{Character ("fwd", "bwd", "tree" or "tfwd") - see \code{fit_graph}}

\item{q}{Penalty term in the stopping criterion (\code{0} = AIC and \code{1} = BIC) - see \code{fit_graph}}

\item{comp}{A list with character vectors. Each elementer in the list is a component in the graph (using expert knowledge)}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{trace}{Logical indicating whether or not to trace the procedure}

\item{validate}{Logical. If true, it checks if \code{A} has only single character values and converts it if not.}

\item{adj}{Adjacency list of a decomposable graph without the observation of interest}
}
\description{
A convenient wrapper for doing outlier test
}
\examples{
\dontrun{
library(dplyr)

# All handwritten digits that have true class equal to a "1".
d <- digits \%>\%
  filter(class == "1") \%>\%
  select(-class)

# A handwritten digit with true class equal to "7"
z <- digits \%>\%
  filter(class == "7") \%>\%
  select(-class) \%>\%
  slice(1) \%>\%
  unlist()

# Test if z is an outlier in class "1"
m <- fit_outlier(as.matrix(d), z)
print(m)
pmf(m)
}
}
\seealso{
\code{\link{outlier_model}}, \code{\link{fit_graph}}
}
