% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gengraph.R
\name{gengraph}
\alias{gengraph}
\title{A generic and extendable structure for decomposable graphical models}
\usage{
gengraph(df, type = "fwd", adj = NULL, q = 0.5, ...)
}
\arguments{
\item{df}{data.frame}

\item{type}{character ("fwd", "bwd", "tree", "tfwd", "gen")}

\item{adj}{A userspecified adjacency list}

\item{q}{Penalty term in the stopping criterion (\code{0} = AIC and \code{1} = BIC)}

\item{...}{Not used (for extendibility)}
}
\value{
A child class of a \code{gengraph} object
}
\description{
A generic structure for decomposable graphical models
}
\examples{
\dontrun{
gengraph(digits, type = "bwd")
gengraph(digits, make_null_graph(colnames(digits)), type = "gen")
}
}
\seealso{
\code{\link{adj_lst.gengraph}}, \code{\link{adj_mat.gengraph}}, \code{\link{fit_graph}}, \code{\link{walk.fwd}}, \code{\link{walk.bwd}}
}
