% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcs}
\alias{mcs}
\title{Maximum Cardinality Search}
\usage{
mcs(adj)
}
\arguments{
\item{adj}{A named adjacency list of a decomposable grah}
}
\value{
A perfect numbering of the nodes
}
\description{
Maximum Cardinality Search
}
