% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcs}
\alias{mcs}
\title{Maximum Cardinality Search}
\usage{
mcs(adj)
}
\arguments{
\item{adj}{A named adjacency list of a decomposable grah}
}
\value{
A list with a perfect numbering of the nodes and a perfect sequence of sets
}
\description{
Maximum Cardinality Search
}
\details{
If adj is not the adjacency list of a decomposable graph and error is produced
}
