% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molic.R
\docType{package}
\name{molic-package}
\alias{molic}
\alias{molic-package}
\title{molic: Multivariate Ouliterdetection In Contingency Tables}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
