% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molic.R
\docType{package}
\name{molic-package}
\alias{molic}
\alias{molic-package}
\title{molic: Multivariate Ouliterdetection In Contingency Tables}
\description{
The molic package has two main functionalities: outlier detection in
contingency tables and model selection in decomposable grahs.
}
\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
