% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{outlier_model}
\alias{outlier_model}
\title{Outlier model}
\usage{
outlier_model(A, adj, nsim = 10000, ncores = 1, validate = TRUE)
}
\arguments{
\item{A}{Character matrix (data)}

\item{adj}{Adjacency list or gengraph object of a decomposable graph without the
observation of interest. See details}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{validate}{Logical. See details.}
}
\description{
A model based on decomposable graphical models for outlier detection
}
\details{
This function is used to find outliers within a dataset. This is different
from \code{fit_outlier} which is used to detect if a new observation is an outlier
in some homogeneous dataset.

It is assumed that all cell values in \code{A}, for all variables,
are represented as a single character. If \code{validate} is \code{TRUE} this is checked.
If cell values are not single characters, one may exploit the \code{to_single_chars} function

The \code{adj} object is most typically found using \code{fit_graph}. But the user can supply
an ajacency list, just a named \code{list}, of their own choice if needed.
}
\examples{
See the 'fit_outlier' procedure
}
\seealso{
\code{\link{fit_outlier}}, \code{\link{fit_graph}}
}
\keyword{internal}
