% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{outlier_model}
\alias{outlier_model}
\title{Outlier model}
\usage{
outlier_model(A, adj, nsim = 5000, ncores = 1, validate = TRUE)
}
\arguments{
\item{A}{Character Matrix (data)}

\item{adj}{Adjacency list of a decomposable graph}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{validate}{Logical. See details.}
}
\description{
A model based on decomposable graphical models for outlier detection
}
\details{
It is assumed that all cell values in \code{A}, for all variables,
are represented as a single character. If \code{validate} is \code{TRUE} this is checked.
If cell values are not single characters, one may exploit the \code{to_single_chars} function
}
\examples{

# All handwritten digits that have true class equal to a "1".
d <- subset(digits, class == "1")
d[, "class"] <- NULL

# A handwritten digit with true class equal to "7"
zd <- subset(digits, class == "7")
z  <- unlist(zd[1, 1:(ncol(zd)-1)])

# Fit an interaction graph
G <- fit_graph(d)
plot(G, vertex.size = 1)

# Append z to the class og "1" digits
dz <- rbind(d, z)

# The outlier model
set.seed(7)
M <- outlier_model(as.matrix(dz), adj_lst(G))
M
pmf(M)

# z is declared as an outlier in the "1" class on a 0.05 significance level
dvz <- deviance(M, z)
pval(M, dvz)
}
