% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{outlier_model}
\alias{outlier_model}
\title{Outlier model}
\usage{
outlier_model(A, adj, nsim = 1000, ncores = 1)
}
\arguments{
\item{A}{Character Matrix (data)}

\item{adj}{Adjacency list of a decomposable graph}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}
}
\description{
A model based on decomposable graphical models for outlier detection
}
\details{
It is assumed that all cell values in \code{A}, for all variables,
are represented as a single character. If \code{validate_A} is \code{TRUE} this is checked.
If cell values are not single characters, one may exploit \code{letters} and \code{LETTERS} e.g.
}
