% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{outlier_model}
\alias{outlier_model}
\title{Outlier tests in contingency tables using decomposable graphical models}
\usage{
outlier_model(df, adj = NULL, nsim = 1000, ncores = 1,
  meta_name = "")
}
\arguments{
\item{df}{Data frame}

\item{adj}{Adjacency list}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{meta_name}{A meta name to keep track of different outlier models}
}
\description{
Outlier tests in contingency tables using decomposable graphical models
}
