% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{outlier_model}
\alias{outlier_model}
\title{Outlier model}
\usage{
outlier_model(A, adj, nsim = 5000, ncores = 1, validate = TRUE)
}
\arguments{
\item{A}{Character Matrix (data) with the new observation of interest appended}

\item{adj}{Adjacency list of a decomposable graph}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{validate}{Logical. See details.}
}
\description{
A model based on decomposable graphical models for outlier detection
}
\details{
It is assumed that all cell values in \code{A}, for all variables,
are represented as a single character. If \code{validate} is \code{TRUE} this is checked.
If cell values are not single characters, one may exploit the \code{to_single_chars} function
}
\examples{
\dontrun{
library(dplyr)

# All handwritten digits that have true class equal to a "1".
d <- digits \%>\% # subset(digits, class == "1")
  filter(class == "1") \%>\%
  select(-class)

# A handwritten digit with true class equal to "7"
z <- digits \%>\%
  filter(class == "7") \%>\%
  select(-class) \%>\%
  slice(1) \%>\%
  unlist()

# Append z to the class of "1" digits
dz <- rbind(d, z)

# Fit an interaction graph
G <- fit_graph(dz, trace = FALSE)
plot(G, vertex.size = 1.5)


# The outlier model
set.seed(7)
m <- outlier_model(as.matrix(dz), adj_lst(G))
print(m)
pmf(m)

# z is declared as an outlier in the "1" class on a 0.05 significance level
dvz <- deviance(m, z)
pval(m, dvz)
}
}
\seealso{
\code{\link{fit_outlier}}, \code{\link{fit_graph}}
}
\keyword{internal}
