% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_model.R
\name{outlier_model}
\alias{outlier_model}
\title{Outlier model}
\usage{
outlier_model(A, adj, nsim = 1000, ncores = 1, validateA = TRUE)
}
\arguments{
\item{A}{Character Matrix (data)}

\item{adj}{Adjacency list of a decomposable graph}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{validateA}{Logical. See details.}
}
\description{
A model based on decomposable graphical models for outlier detection
}
\details{
It is assumed that all cell values in \code{A}, for all variables,
are represented as a single character. If \code{validate_A} is \code{TRUE} this is checked.
If cell values are not single characters, one may exploit \code{letters} and \code{LETTERS} e.g.
}
\examples{


# DNA data for Europeans
df   <- subset(tgp_dat[, c(2, 5:10)], pop_meta == "EUR")                

# DNA profile from Africa
z    <- unlist(subset(tgp_dat[, c(2, 5:10)], pop_meta == "AFR")[1, -1])

# Fit an interactiongraph
G    <- adj_list(efs(df[, -1]))

# Append z to df
df_z <- rbind(df[, -1], z)

# The outlier model
set.seed(7)
M    <- outlier_model(as.matrix(df_z), G)

# z is declared as an outlier in the Europe data on a 0.05 significance level 
p_val(M, deviance(M, z))

# Plot the distribution of the deviance
pmf(M)

}
