% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{outliers}
\alias{outliers}
\title{Detect Outliers}
\usage{
outliers(x, d, alpha = 0.05)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{d}{Character matrix (data)}

\item{alpha}{Sigficance level}
}
\description{
This function finds the outliers in some data given an outlier model
}
