% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{plot.gengraph}
\alias{plot.gengraph}
\title{Plot}
\usage{
\method{plot}{gengraph}(x, vc = NULL, ...)
}
\arguments{
\item{x}{A \code{gengraph} object}

\item{vc}{Named character vector; the names are the vertices and
the elements are the colors of the nodes}

\item{...}{Extra arguments. See the igraph package}
}
\description{
A wrapper around igraphs plot method for \code{gengraph} objects
}
\examples{
d  <- subset(digits, class == "1")[, 20:30]
g  <- fit_graph(d)
vs <- names(g$G_adj)
vcol <- structure(vector("character", length(vs)), names = names(g$G_adj))
vcol[1:4]  <- "lightsteelblue2"
vcol[5:7]  <- "orange"
vcol[8:11] <- "pink"
plot(g, vcol)
}
