% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{plot.multiple_models}
\alias{plot.multiple_models}
\title{Plot Deviance of Multiple Models}
\usage{
\method{plot}{multiple_models}(x, sig_col = "#FF0000A0", ...)
}
\arguments{
\item{x}{A \code{multiple_models} object returned from a called to \code{fit_multiple_models}}

\item{sig_col}{Color of the significance level area (default is red)}

\item{...}{Extra arguments. See details.}
}
\value{
No return value, called for side effects
}
\description{
A plot method to show the approximated deviance distribution of multiple models
}
\details{
The dotted line represents the observed deviance of the observation under the hypothesis
and the colored (red is default) area under the graph represents the significance level.
Thus, if the dotted line is to the left of the colored area, the hypothesis that the observation
is an outlier cannot be rejected. Notice however, if there is no dotted line, this simply means,
that the observed deviance is larger than all values and it would disturb the plot if included.

No extra arguments \code{...} are implement at the moment.
}
