% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{plot.outlier_model}
\alias{plot.outlier_model}
\title{Plot of pmf}
\usage{
\method{plot}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatibility)}
}
\description{
A plot method to show the pmf of the approximated pmf of \code{T(Y)}
}
