% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{print.csptable}
\alias{print.csptable}
\title{Print method for conditional sparse tables}
\usage{
\method{print}{csptable}(x, ...)
}
\arguments{
\item{x}{A \code{csptable} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
Print method for conditional sparse tables
}
