% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{print.sptable}
\alias{print.sptable}
\title{Print Sparse Table}
\usage{
\method{print}{sptable}(x, ...)
}
\arguments{
\item{x}{A \code{sptable} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{sptable} and \code{slice_sptable} objects
}
