% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tables.R
\name{sptable}
\alias{sptable}
\title{Sparse table}
\usage{
sptable(x)
}
\arguments{
\item{x}{Character matrix}
}
\description{
Returns a sparse contingency table for the variables in \code{x} as a vector .
}
\examples{
sptable(as.matrix(tgp_dat[, 5:8]))
}
\seealso{
\code{\link{csptable}}
}
