% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TY}
\alias{TY}
\title{\code{T(y)}}
\usage{
TY(y, C_marginals, S_marginals)
}
\arguments{
\item{y}{A named vector (named according to data)}

\item{C_marginals}{Clique marginal tables (use a_marginals function)}

\item{S_marginals}{Separator marginal tables (use a_marginals function)}
}
\description{
This function Calculates the affine value \code{T(y)} of \code{-2\log} likelihood-ratio statistic
}
