% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{variance}
\alias{variance}
\alias{variance.outlier_model}
\title{Variance}
\usage{
variance(x)

\method{variance}{outlier_model}(x, ...)
}
\arguments{
\item{x}{A \code{outlier_model} object}

\item{...}{Not used (for S3 compatibility)}
}
\description{
Estimated variance of the deviance statistic \code{T(Y)}
}
